+++ src/main/java/tornadofx/control/DateTimePicker.java
@@ -96,7 +96,7 @@ public class DateTimePicker extends DatePicker {
 		}
 
 		public LocalDate fromString(String value) {
-			if (value == null) {
+			if (value == null || value.isEmpty()) {
 				dateTimeValue.set(null);
 				return null;
 			}