+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Change Log
-All notable changes to this project will be documented in this file.
-## [Unreleased]
+
+## [1.0.5]
+
+- - Changed NaviSelect to render as dropdown with button, the previous look confused people into thinking it was editable as a text field
+- - Use correct submit event for DateTimePicker
 
 ### Changed